pyproject.toml
[project]
name = "tts-pro-main"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.129.0",
    "gtts>=2.5.4",
    "openai>=2.21.0",
    "passlib[bcrypt]>=1.7.4",
    "pydantic>=2.12.5",
    "pydub>=0.25.1",
    "python-dotenv>=1.2.1",
    "python-jose>=3.5.0",
    "python-multipart>=0.0.22",
    "reportlab>=4.4.10",
    "slowapi>=0.1.9",
    "sqlalchemy>=2.0.46",
    "uvicorn>=0.41.0",
]
